+++ examples/examples_nodejs_test.go
@@ -19,6 +19,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	appsv1 "k8s.io/api/apps/v1"
 	"os"
 	"os/exec"
 	"path"
@@ -31,16 +32,35 @@ import (
 )
 
 func TestAccCluster(t *testing.T) {
+
 	test := getJSBaseOptions(t).
 		With(integration.ProgramTestOptions{
 			Dir:           path.Join(getCwd(t), "./cluster"),
-			RunUpdateTest: true,
+			RunUpdateTest: false,
 			ExtraRuntimeValidation: func(t *testing.T, info integration.RuntimeValidationStackInfo) {
 				utils.RunEKSSmokeTest(t,
 					info.Deployment.Resources,
 					info.Outputs["kubeconfig1"],
 					info.Outputs["kubeconfig2"],
 				)
+
+				assert.NoError(t, utils.ValidateDaemonSet(t, info.Outputs["kubeconfig2"], "kube-system", "aws-node", func(ds *appsv1.DaemonSet) {
+					for _, c := range ds.Spec.Template.Spec.Containers {
+						assert.Equal(t, "602401143452.dkr.ecr.us-west-2.amazonaws.com/amazon-k8s-cni:v1.9.0", c.Image)
+					}
+
+					for _, ic := range ds.Spec.Template.Spec.InitContainers {
+						assert.Equal(t, "602401143452.dkr.ecr.us-west-2.amazonaws.com/amazon-k8s-cni-init:v1.9.0", ic.Image)
+						var tcpEarly bool
+						for _, env := range ic.Env {
+							if env.Name == "DISABLE_TCP_EARLY_DEMUX" {
+								tcpEarly = env.Value == "true"
+								assert.Equal(t, env.Value, "true")
+							}
+						}
+						assert.True(t, tcpEarly)
+					}
+				}))
 			},
 		})