+++ nodejs/eks/cni.ts
@@ -80,6 +80,13 @@ export interface VpcCniOptions {
     image?: pulumi.Input<string>;
 
     /**
+     * Specifies the init container image to use in the AWS CNI cluster DaemonSet.
+     *
+     * Defaults to the official AWS CNI init container image in ECR.
+     */
+    initImage?: pulumi.Input<string>;
+
+    /**
      * Specifies the veth prefix used to generate the host-side veth device
      * name for the CNI.
      *
@@ -123,9 +130,19 @@ export interface VpcCniOptions {
      * Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label to the node if the
      * instance has capacity to attach an additional ENI.
      *
+     * If using liveness and readiness probes, you will also need to disable TCP early demux.
+     *
      * Defaults to "false".
      */
     enablePodEni?: pulumi.Input<boolean>;
+
+    /**
+     * Allows the kubelet's liveness and readiness probes to connect via TCP when pod ENI is enabled.
+     * This will slightly increase local TCP connection latency.
+     *
+     * Defaults to "false".
+     */
+    disableTcpEarlyDemux?: pulumi.Input<boolean>;
 
     /**
      * Specifies whether ipamd should configure rp filter for primary interface.
@@ -194,13 +211,15 @@ export class VpcCni extends pulumi.CustomResource {
             logLevel: args?.logLevel,
             logFile: args?.logFile,
             image: args?.image,
+            initImage: args?.initImage,
             eniConfigLabelDef: args?.eniConfigLabelDef,
             pluginLogLevel: args?.pluginLogLevel,
             pluginLogFile: args?.pluginLogFile,
             enablePodEni: args?.enablePodEni,
+            disableTcpEarlyDemux: args?.disableTcpEarlyDemux,
             cniConfigureRpfilter: args?.cniConfigureRpfilter,
             cniCustomNetworkCfg: args?.cniCustomNetworkCfg,
-            cniExternalSnat: args?.cniCustomNetworkCfg,
+            cniExternalSnat: args?.cniExternalSnat,
             securityContextPrivileged: args?.securityContextPrivileged,
         }, opts);
     }