+++ sdk/go/eks/pulumiTypes.go
@@ -2354,7 +2354,9 @@ type VpcCniOptions struct {
 	//
 	// Defaults to false.
 	CustomNetworkConfig *bool `pulumi:"customNetworkConfig"`
-	// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label tothe node if the instance has capacity to attach an additional ENI. Default is `false`.
+	// Allows the kubelet's liveness and readiness probes to connect via TCP when pod ENI is enabled. This will slightly increase local TCP connection latency.
+	DisableTcpEarlyDemux *bool `pulumi:"disableTcpEarlyDemux"`
+	// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label to the node if the instance has capacity to attach an additional ENI. Default is `false`. If using liveness and readiness probes, you will also need to disable TCP early demux.
 	EnablePodEni *bool `pulumi:"enablePodEni"`
 	// Specifies the ENI_CONFIG_LABEL_DEF environment variable value for worker nodes. This is used to tell Kubernetes to automatically apply the ENIConfig for each Availability Zone
 	// Ref: https://docs.aws.amazon.com/eks/latest/userguide/cni-custom-network.html (step 5(c))
@@ -2373,6 +2375,10 @@ type VpcCniOptions struct {
 	//
 	// Defaults to the official AWS CNI image in ECR.
 	Image *string `pulumi:"image"`
+	// Specifies the init container image to use in the AWS CNI cluster DaemonSet.
+	//
+	// Defaults to the official AWS CNI init container image in ECR.
+	InitImage *string `pulumi:"initImage"`
 	// Specifies the file path used for logs.
 	//
 	// Defaults to "stdout" to emit Pod logs for `kubectl logs`.
@@ -2425,7 +2431,9 @@ type VpcCniOptionsArgs struct {
 	//
 	// Defaults to false.
 	CustomNetworkConfig pulumi.BoolPtrInput `pulumi:"customNetworkConfig"`
-	// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label tothe node if the instance has capacity to attach an additional ENI. Default is `false`.
+	// Allows the kubelet's liveness and readiness probes to connect via TCP when pod ENI is enabled. This will slightly increase local TCP connection latency.
+	DisableTcpEarlyDemux pulumi.BoolPtrInput `pulumi:"disableTcpEarlyDemux"`
+	// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label to the node if the instance has capacity to attach an additional ENI. Default is `false`. If using liveness and readiness probes, you will also need to disable TCP early demux.
 	EnablePodEni pulumi.BoolPtrInput `pulumi:"enablePodEni"`
 	// Specifies the ENI_CONFIG_LABEL_DEF environment variable value for worker nodes. This is used to tell Kubernetes to automatically apply the ENIConfig for each Availability Zone
 	// Ref: https://docs.aws.amazon.com/eks/latest/userguide/cni-custom-network.html (step 5(c))
@@ -2444,6 +2452,10 @@ type VpcCniOptionsArgs struct {
 	//
 	// Defaults to the official AWS CNI image in ECR.
 	Image pulumi.StringPtrInput `pulumi:"image"`
+	// Specifies the init container image to use in the AWS CNI cluster DaemonSet.
+	//
+	// Defaults to the official AWS CNI init container image in ECR.
+	InitImage pulumi.StringPtrInput `pulumi:"initImage"`
 	// Specifies the file path used for logs.
 	//
 	// Defaults to "stdout" to emit Pod logs for `kubectl logs`.
@@ -2573,7 +2585,12 @@ func (o VpcCniOptionsOutput) CustomNetworkConfig() pulumi.BoolPtrOutput {
 	return o.ApplyT(func(v VpcCniOptions) *bool { return v.CustomNetworkConfig }).(pulumi.BoolPtrOutput)
 }
 
-// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label tothe node if the instance has capacity to attach an additional ENI. Default is `false`.
+// Allows the kubelet's liveness and readiness probes to connect via TCP when pod ENI is enabled. This will slightly increase local TCP connection latency.
+func (o VpcCniOptionsOutput) DisableTcpEarlyDemux() pulumi.BoolPtrOutput {
+	return o.ApplyT(func(v VpcCniOptions) *bool { return v.DisableTcpEarlyDemux }).(pulumi.BoolPtrOutput)
+}
+
+// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label to the node if the instance has capacity to attach an additional ENI. Default is `false`. If using liveness and readiness probes, you will also need to disable TCP early demux.
 func (o VpcCniOptionsOutput) EnablePodEni() pulumi.BoolPtrOutput {
 	return o.ApplyT(func(v VpcCniOptions) *bool { return v.EnablePodEni }).(pulumi.BoolPtrOutput)
 }
@@ -2605,6 +2622,13 @@ func (o VpcCniOptionsOutput) ExternalSnat() pulumi.BoolPtrOutput {
 // Defaults to the official AWS CNI image in ECR.
 func (o VpcCniOptionsOutput) Image() pulumi.StringPtrOutput {
 	return o.ApplyT(func(v VpcCniOptions) *string { return v.Image }).(pulumi.StringPtrOutput)
+}
+
+// Specifies the init container image to use in the AWS CNI cluster DaemonSet.
+//
+// Defaults to the official AWS CNI init container image in ECR.
+func (o VpcCniOptionsOutput) InitImage() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v VpcCniOptions) *string { return v.InitImage }).(pulumi.StringPtrOutput)
 }
 
 // Specifies the file path used for logs.
@@ -2721,7 +2745,17 @@ func (o VpcCniOptionsPtrOutput) CustomNetworkConfig() pulumi.BoolPtrOutput {
 	}).(pulumi.BoolPtrOutput)
 }
 
-// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label tothe node if the instance has capacity to attach an additional ENI. Default is `false`.
+// Allows the kubelet's liveness and readiness probes to connect via TCP when pod ENI is enabled. This will slightly increase local TCP connection latency.
+func (o VpcCniOptionsPtrOutput) DisableTcpEarlyDemux() pulumi.BoolPtrOutput {
+	return o.ApplyT(func(v *VpcCniOptions) *bool {
+		if v == nil {
+			return nil
+		}
+		return v.DisableTcpEarlyDemux
+	}).(pulumi.BoolPtrOutput)
+}
+
+// Specifies whether to allow IPAMD to add the `vpc.amazonaws.com/has-trunk-attached` label to the node if the instance has capacity to attach an additional ENI. Default is `false`. If using liveness and readiness probes, you will also need to disable TCP early demux.
 func (o VpcCniOptionsPtrOutput) EnablePodEni() pulumi.BoolPtrOutput {
 	return o.ApplyT(func(v *VpcCniOptions) *bool {
 		if v == nil {
@@ -2777,6 +2811,18 @@ func (o VpcCniOptionsPtrOutput) Image() pulumi.StringPtrOutput {
 			return nil
 		}
 		return v.Image
+	}).(pulumi.StringPtrOutput)
+}
+
+// Specifies the init container image to use in the AWS CNI cluster DaemonSet.
+//
+// Defaults to the official AWS CNI init container image in ECR.
+func (o VpcCniOptionsPtrOutput) InitImage() pulumi.StringPtrOutput {
+	return o.ApplyT(func(v *VpcCniOptions) *string {
+		if v == nil {
+			return nil
+		}
+		return v.InitImage
 	}).(pulumi.StringPtrOutput)
 }